feat: make the space use params
