+++ src/app/(authenticated)/space/[spaceId]/page.tsx
