+++ src/components/global/sidebar/nav-spaces.tsx
@@ -7,7 +7,6 @@ import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
@@ -22,10 +21,12 @@ import type { ExtendedUser } from "../../../../next-auth";
 
 import { useRouter } from "next/navigation";
 import { useUserSpaces } from "@/queries/spaces";
+import { useSpaceId } from "@/hooks/useSpaceId";
 
 export function NavSpaces({ user }: { user: Pick<ExtendedUser, "id"> }) {
   const { isMobile } = useSidebar();
   const router = useRouter();
+  const spaceId = useSpaceId();
 
   const { data: spaces = [], isLoading, error } = useUserSpaces(user?.id || "");
 
@@ -39,8 +40,14 @@ export function NavSpaces({ user }: { user: Pick<ExtendedUser, "id"> }) {
   }
 
   const handleCreateSpace = async () => {
-    router.push("/spaces");
+    router.push("/create-space");
   };
+
+  const currentSpace = spaces.find((space) => space.id === spaceId);
+
+  const filteredSpaces = spaces?.filter(
+    (space) => space.id !== currentSpace?.id,
+  );
 
   return (
     <SidebarMenu>
@@ -59,8 +66,12 @@ export function NavSpaces({ user }: { user: Pick<ExtendedUser, "id"> }) {
                 <AvatarFallback className="rounded-lg">CN</AvatarFallback>
               </Avatar>
               <div className="grid flex-1 text-left text-sm leading-tight">
-                <span className="truncate font-semibold">name</span>
-                <span className="truncate text-xs">id</span>
+                <span className="truncate font-semibold">
+                  {currentSpace?.name}
+                </span>
+                <span className="truncate text-xs">
+                  {currentSpace?.description}
+                </span>
               </div>
               <ChevronsUpDown className="ml-auto size-4" />
             </SidebarMenuButton>
@@ -71,28 +82,25 @@ export function NavSpaces({ user }: { user: Pick<ExtendedUser, "id"> }) {
             align="end"
             sideOffset={4}
           >
-            <DropdownMenuLabel className="p-0 font-normal">
-              {spaces.map((space) => (
-                <div
-                  key={space.id}
-                  className="flex items-center gap-2 px-1 py-1.5 text-left text-sm"
-                >
-                  <Avatar className="h-8 w-8 rounded-lg">
-                    <AvatarImage
-                      src={user.id || undefined}
-                      alt={user.id || "User"}
-                    />
-                    <AvatarFallback className="rounded-lg">CN</AvatarFallback>
-                  </Avatar>
-                  <div className="grid flex-1 text-left text-sm leading-tight">
-                    <span className="truncate font-semibold">{space.name}</span>
-                    <span className="truncate text-xs">
-                      {space.description}
-                    </span>
-                  </div>
+            {filteredSpaces.map((space) => (
+              <DropdownMenuItem
+                key={space.id}
+                onClick={() => router.push(`/space/${space.id}`)}
+                className="flex items-center gap-2 px-1 py-1.5 text-left text-sm"
+              >
+                <Avatar className="h-8 w-8 rounded-lg">
+                  <AvatarImage
+                    src={user.id || undefined}
+                    alt={user.id || "User"}
+                  />
+                  <AvatarFallback className="rounded-lg">CN</AvatarFallback>
+                </Avatar>
+                <div className="grid flex-1 text-left text-sm leading-tight">
+                  <span className="truncate font-semibold">{space.name}</span>
+                  <span className="truncate text-xs">{space.description}</span>
                 </div>
-              ))}
-            </DropdownMenuLabel>
+              </DropdownMenuItem>
+            ))}
             <DropdownMenuSeparator />
 
             <DropdownMenuItem onClick={handleCreateSpace}>
