+++ src/routes.ts
@@ -29,4 +29,4 @@ export const apiAuthPrefix = "/api/auth";
  * The default redirect path after loggin in
  * @type {string}
  */
-export const DEFAULT_LOGIN_REDIRECT = "/home";
+export const DEFAULT_LOGIN_REDIRECT = "/space";
